+++ mas-foundation/src/main/java/com/ca/mas/foundation/MASCallback.java
@@ -10,25 +10,32 @@ package com.ca.mas.foundation;
 
 import android.os.Handler;
 
-public abstract class MASCallback<T> {
+/**
+ * Generic callback interface
+ *
+ * @param <T> The object type that will be received when {@link #onSuccess(Object)} is invoked.
+ */
+public interface MASCallback<T> {
 
     /**
      * The Handler to handle the callback, refer to {@link Handler} for details
      */
-    public Handler getHandler() {
+    default Handler getHandler() {
         return null;
     }
 
     /**
      * Called when an asynchronous call completes successfully.
+     *
      * @param result the value returned
      */
-    public abstract void onSuccess(T result);
+    void onSuccess(T result);
 
     /**
      * Called when an asynchronous call fails to complete.
+     *
      * @param e the reason for failure
      */
-    public abstract void onError(Throwable e);
+    void onError(Throwable e);
 
 }