+++ mas-test/src/main/java/com/ca/mas/MASCallbackFuture.java
@@ -22,7 +22,7 @@ import java.util.concurrent.TimeoutException;
  * A Future represents the result of the MASCallback asynchronous result.
  * @param <T> The result type returned by this Future's get method
  */
-public class MASCallbackFuture<T> extends MASCallback<T> implements Future<T> {
+public class MASCallbackFuture<T> implements MASCallback<T> ,Future<T> {
 
     private CountDownLatch countDownLatch = new CountDownLatch(1);
     private boolean done = false;